Добавлена возможность отмены добавления и редактирования. Задачи отсортированы по дате

+++ main.py
@@ -15,7 +15,7 @@ def helpdesk():
         tomorrow = today + datetime.timedelta(days=1)
         clear_scope('output')
         output_list, interaction_list = [], []
-        task_list = db.get_tasks(False)
+        task_list = db.get_tasks(False, "date")
         number = 1
         for task in task_list:
             if task[4] == today:
@@ -53,35 +53,39 @@ def helpdesk():
                 textarea('Текст задания:', name='description'),
                 input('Кто обратился?', type=TEXT, name='responsible'),
                 input('Какой кабинет?', type=TEXT, name='cabinet'),
-                input('Укажите крайний срок выполнения', type=DATE, name='date'),
+                input('Укажите крайний срок выполнения', type=DATE, name='date', value=str(today)),
                 input('Комментарий к заданию', type=TEXT, name='comment')
-            ])
+            ], cancelable=True)
             
-            db.set_task([
-                new_task['description'],
-                new_task['responsible'],
-                new_task['cabinet'],
-                new_task["date"],
-                new_task['comment']
-                ])
+            if new_task != None:
+                db.set_task([
+                    new_task['description'],
+                    new_task['responsible'],
+                    new_task['cabinet'],
+                    new_task["date"],
+                    new_task['comment']
+                    ])
             
         elif menu == 'delete_task':
             num = input('Введите номер задачи для удаления:', type=NUMBER)
-            for element in interaction_list:
-                if element[0] == num:
-                    id = element[1]
-            db.delete_task(id)
+            if num != None:
+                for element in interaction_list:
+                    if element[0] == num:
+                        id = element[1]
+                db.delete_task(id)
         
         elif menu == 'mark_execution':
             num = input('Введите номер задачи чтобы пометить её как завершенное:', type=NUMBER)
-            for element in interaction_list:
-                if element[0] == num:
-                    id = element[1]
-            db.set_status_ready(id)
+            if num != None:
+                for element in interaction_list:
+                    if element[0] == num:
+                        id = element[1]
+                db.set_status_ready(id)
 
         elif menu == 'get_ready_tasks':
             output_list = []
-            task_ready_list = db.get_tasks(True)
+            task_ready_list = db.get_tasks(True, "completed")
+            task_ready_list.reverse()
             number = 1
             for task in task_ready_list:
                 output_list.append((number, task[1], task[2], task[3], task[5], task[7].strftime("%d.%m.%Y")))
@@ -100,32 +104,35 @@ def helpdesk():
 
         elif menu == 'edit_task':
             num = input('Введите номер задачи, которую нужно отредактировать:', type=NUMBER)
-            for element in interaction_list:
-                if element[0] == num:
-                    id = element[1]
-            data_task = db.get_tasks(False)
-            for task in data_task:
-                if task[0] == id:
-                    data = task
-            clear_scope('output')
-            edit_task = input_group('Добавить задачу', [
-                textarea('Текст задания:', name='description', value=data[1]),
-                input('Кто обратился?', type=TEXT, name='responsible', value=data[2]),
-                input('Какой кабинет?', type=TEXT, name='cabinet', value=data[3]),
-                input('Укажите крайний срок выполнения', type=DATE, name='date', value=data[4].strftime("%Y-%m-%d")),
-                input('Комментарий к заданию', type=TEXT, name='comment', value=data[5])
-            ])
-
-            data_list = [
-                data[0],
-                edit_task['description'],
-                edit_task['responsible'],
-                edit_task['cabinet'],
-                edit_task['date'],
-                edit_task['comment'] 
-                ]
             
-            db.update_task(data_list)
+            if num != None:
+                for element in interaction_list:
+                    if element[0] == num:
+                        id = element[1]
+                data_task = db.get_tasks(False, "date")
+                for task in data_task:
+                    if task[0] == id:
+                        data = task
+                clear_scope('output')
+                edit_task = input_group('Добавить задачу', [
+                    textarea('Текст задания:', name='description', value=data[1]),
+                    input('Кто обратился?', type=TEXT, name='responsible', value=data[2]),
+                    input('Какой кабинет?', type=TEXT, name='cabinet', value=data[3]),
+                    input('Укажите крайний срок выполнения', type=DATE, name='date', value=data[4].strftime("%Y-%m-%d")),
+                    input('Комментарий к заданию', type=TEXT, name='comment', value=data[5])
+                ], cancelable=True)
+            
+                if edit_task != None:
+                    data_list = [
+                        data[0],
+                        edit_task['description'],
+                        edit_task['responsible'],
+                        edit_task['cabinet'],
+                        edit_task['date'],
+                        edit_task['comment'] 
+                        ]
+                    
+                    db.update_task(data_list)
 
 
 start_server(helpdesk, port=8000, remote_access=True, auto_open_webbrowser=True, debug=True)
